server/requirements.txt
websockets
glicko2
python-dotenv